requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
tweepy==4.14.0
transformers==4.35.0
torch==2.8.0
smtplib  # Built-in, but for clarity